+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   Upgraded to Svelte 4, and associated dependencies.
 -   Fixed several missing ARIA roles on interactive elements
+-   Fixed ConfirmButton, which didn't revert to non-confirming state after evaluating command.
 
 # 0.5.32 - 2023-06-17